Structure of \iReal_tree changed, update of excel classification sheet
We now seperate the csv files in the \iReal_tree folder, depending on whether they are finished or still being worked on. Finished csv files have to be put in the "finished_trees" folder.

Then, the classification sheet located in the folder "iRealPro", above the "iReal_trees" folder, can search through the files located in these folders and update in the sheet which tunes are being worked on and which are already finished