interplay, birdkswork and classification file
hi!

During early july i worked on the classification file, providing a small macro that based on the architecture of the folders, sorts which standards are being worked on or already analysed.

Then, I added the sorting for those which start with a dominant chord

And analysed few other standards..but I have many things to do here in rio for my study on choro and brazilian music, so I am way outdated